requirements.txt
Django==2.1.3
django-geojson==2.12.0
django-leaflet==0.24.0
django-location-field==2.0.6
django-rest-framework==0.1.0
djangorestframework==3.9.0
gunicorn==19.9.0
jsonfield==2.0.2
psycopg2==2.7.6.1
pytz==2018.7
six==1.12.0
